elo lb pages, elo lb placeholders, elo tier perks

+++ src/main/java/com/renatusnetwork/parkour/data/menus/MenuItemAction.java
@@ -346,7 +346,7 @@ public class MenuItemAction {
 
                 if (!event.isCancelled())
                 {
-                    int playerBalance = (int) playerStats.getCoins();
+                    int playerBalance = playerStats.getCoins();
                     int price = event.getPrice();
 
                     if (playerStats.hasModifier(ModifierType.SHOP_DISCOUNT))
